+++ CHANGELOG.md
@@ -3,19 +3,26 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 # [v1.7.0](Unreleased)
 ### Notes
-- This release extends supports of the SDK to Oneview API Version 2200 and Image Streamer API Version 2000.
+- This release extends supports of the SDK to Oneview API2200.
 
 ### Oneview Features supported
 - Deployment Plan
+- Enclosure
+- Enclosure Group
 - Hypervisor Cluster Profiles
 - Hypervisor Manager
 - Interconnects
 - Interconnect Types
 - Logical Interconnect
+- Logical Enclosure
+- Logical Interconnect Group
+- Scope
 - Server Certificates
 - Server Hardware
 - Server Hardware Types
 - Storage Volume Attachments
+- Tasks
+- Uplink Set
 
 # [v1.6.0] 
 ### Notes